extract verifying a signed object from the device tracker
